feat: 항공사 최소값 리스트 동적 렌더링
